Create enhanced LimitDialog
Now it uses the NXPanel so that limits and other plot parameters can be copied from other tabs in the panel. This also converts the projection panels to use NXPanel. These were formerly defined in plotview.py.
